+++ src/orbitmesh-console/Pages/PackageRepository.js
@@ -171,7 +171,7 @@ export default {
 
             <table class="data-table">
                 <thead>
-                    <tr><th style="width:140px">Name</th><th style="width:80px">Version</th><th style="width:70px">Runtime</th><th>Description</th><th style="width:90px">Feed</th><th style="width:160px">File</th><th style="width:90px">Size</th><th style="width:170px">Last update</th><th style="width:340px">Actions</th></tr>
+                    <tr><th style="width:140px">Name</th><th style="width:80px">Version</th><th style="width:80px">Runtime</th><th>Description</th><th style="width:110px">Feed</th><th style="width:160px">File</th><th style="width:90px">Size</th><th style="width:170px">Last update</th><th style="width:340px">Actions</th></tr>
                 </thead>
                 <tbody>
                     <tr v-for="pkg in (store.packagesRepository || [])" :key="pkg.Filename">